+++ server/src/task/test/data/index.ts
@@ -1,0 +1,8 @@
+export {
+  invalidCreateTaskFields,
+  validCreateTaskField,
+} from './createTaskData';
+export {
+  invalidUpdateTaskFields,
+  validUpdateTaskField,
+} from './updateTaskData';
